COMPUTE_PIPELINE v0.33.1: CI rot nach f068cb4; die erzeugte Modellkarte war alt, und clippy fand elf   Stellen, davon zwei echte Maengel: ein verwaister Doc-Kommentar an der falschen Funktion, eine   352-Byte-Variante | ETHICS/Modellkarte.md (neu erzeugt, theta_v 0.18.0);   COMPUTE_PIPELINE/myl-pod/src/{shardweg,trainingswerk}.rs und tests (zwei);   INTEGER_LLM/runtime/src/{trainingsschleife,shardtraining,messung}.rs;   INTEGER_LLM/kernels/src/{konformitaet,optimierer}.rs; CONSENSUS/myl-ledger/src/transitions.rs;   NODE/myl-node/src/stichprobe.rs; TESTCLIENT/myl-testclient/tests/verzahnte_simulation.rs; AGENTS.md   (clippy als Pflichtschritt vor dem Commit); Cargo.toml, README/README.md

+++ COMPUTE_PIPELINE/myl-pod/src/trainingswerk.rs
@@ -169,7 +169,7 @@ impl Trainingswerk {
         for s in 0..v.schritte {
             // 1. Einbettung und Vorwärtslauf durch alle Shards.
             let mut strom: Vec<Vec<i16>> =
-                v.folge.iter().map(|t| m.embed_token(*t as usize)).collect();
+                v.folge.iter().map(|t| m.embed_token(*t)).collect();
             let mut mitschnitte = Vec::with_capacity(SHARDS);
             for j in 0..SHARDS {
                 let vg = self.vorgaben(j, s, v);
